requirements.txt
# Core ML/DL
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Visualization (for notebooks)
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Load Testing
locust==2.18.3

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
